Move network Result state object to util package.
