+++ app/src/main/java/ru/sandbox/androidacademyapp/repository/IMovieRepository.kt
@@ -1,6 +1,6 @@
 package ru.sandbox.androidacademyapp.repository
 
-import ru.sandbox.androidacademyapp.data.network.Result
+import ru.sandbox.androidacademyapp.util.Result
 import ru.sandbox.androidacademyapp.data.db.entites.Movie
 import ru.sandbox.androidacademyapp.data.db.entites.relations.MovieWithActors
