+++ app/src/main/java/ru/sandbox/androidacademyapp/repository/MovieRepository.kt
@@ -6,7 +6,7 @@ import ru.sandbox.androidacademyapp.BuildConfig
 import ru.sandbox.androidacademyapp.data.network.MoviesApi
 import ru.sandbox.androidacademyapp.data.network.ActorResponse
 import ru.sandbox.androidacademyapp.data.network.MovieResponse
-import ru.sandbox.androidacademyapp.data.network.Result
+import ru.sandbox.androidacademyapp.util.Result
 import ru.sandbox.androidacademyapp.data.db.MoviesDao
 import ru.sandbox.androidacademyapp.data.db.entites.Actor
 import ru.sandbox.androidacademyapp.data.db.entites.Movie
